Added new HTML page, CSS styling and SVG logo
A new HTML page has been created with basic structure including meta tags, title, stylesheet link, body content and a script. The body contains an image tag for the logo, headings, paragraphs with contact information and links to support resources. A footer is also included with company details.

In addition to this, a new CSS file has been added which includes styles for the body text alignment and font family as well as fixed positioning and styling for the footer.

Lastly, an SVG logo has been added in the img directory. This logo is referenced in the newly created HTML file.